LayoutTests/http/tests/security/xss-ALLOWED-xsl-external-entity-xslt-docloader-expected.txt

ALERT: document() This test passes if this message was loaded.
